Add files via upload
For both Dog or Cat CNN and Traffic Sign CNN I couldn't; upload datasets, but for a dog or cat CNN we had 4000 pics of both dogs and cats for training data and 1000 for both dogs and cats for test data, and for traffic sign, CNN we had 43000+ pics for training data, 12000+ pics for test data and 6000+ pics for validation data with 43 classes of pics like it was written at the beginning of the code. For CIFAR-10 I went low on epochs and batch size and also with artificial neurons because even with these numbers it took a long time I inserted these numbers into my model so I could finish the code but if anyone of you guys wants to experiment with numbers feel free because I bet there is a better combination of numbers to make the model better!